refactor(user-list): move classes in the appropriate packages
